Implement preserving of book and label classes
