+++ store/books.json
@@ -1,1 +1,1 @@
-[]
+[{"publisher":"Yonas","cover_state":"bad","publish_date":"today","label":"Gift"}]
