+++ store/labels.json
@@ -1,1 +1,1 @@
-[]
+[{"title":"Gift","color":"Black","items":[{"publisher":"Yonas","cover_state":"bad","publish_date":"today","label":"Gift"}]}]
